common code extracted
