+++ src/storefront/pages/order/[orderId]/index.tsx
@@ -1,15 +1,9 @@
 import { GetServerSideProps, NextApiRequest, NextApiResponse } from "next";
-import { useRouter } from "next/router";
 import getConfig from "next/config";
 import React from "react";
-// utils
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { useTranslation } from "next-i18next";
 import { cartBillingInfoAPI } from "@/pages/api/cart/[token]/billing-info";
 import { cartShippingInfoAPI } from "@/pages/api/cart/[token]/shipping-info";
-import CartStepper from "@/components/Cart/Stepper";
-import { Box, Table, Typography } from "@mui/material";
-import Grid from "@mui/material/Grid";
 import {
   IBillingInfo,
   IPaymentMethodCountry,
@@ -18,23 +12,16 @@ import {
 } from "@/types/cart";
 import { countryListAPI } from "@/pages/api/country";
 import { ICountry } from "@/types/country";
-import CartButtonRow from "@/components/Cart/ButtonRow";
-import CartItemList from "@/components/Cart/CartItemList";
-import CartInfoSummary, {
-  ICartInfoTableRow,
-} from "@/components/Cart/CartInfoSummary";
-import CartMethodSummaryInfoRow from "@/components/Cart/Methods/CartMethodSummaryInfoRow";
 import { cartPaymentMethodAPI } from "@/pages/api/cart/[token]/payment-method";
 import { cartShippingMethodAPI } from "@/pages/api/cart/[token]/shipping-method";
-import { useTheme } from "@mui/material/styles";
-import Divider from "@mui/material/Divider";
-import { useCart } from "@/utils/context/cart";
-import CollapsableContentWithTitle from "@/components/Generic/CollapsableContentWithTitle";
-import CartCompleteOrder from "@/components/Cart/CartCompleteOrder";
+import CartOrderSummary from "@/components/Cart/CartOrderSummary";
+import { IOrderWithPayment } from "@/types/order";
+import { orderDetailAPI } from "@/pages/api/order/[id]";
 
 const { serverRuntimeConfig } = getConfig();
 
 interface ICartSummaryPageProps {
+  orderWithPayment: IOrderWithPayment;
   billingInfo: IBillingInfo;
   shippingInfo: IShippingInfo;
   countries: ICountry[];
@@ -45,158 +32,26 @@ interface ICartSummaryPageProps {
 /**
  * Cart summary page displaying all the items, addresses and shipping & payment method
  */
-const CartSummaryPage = ({
+const OrderDetailPage = ({
+  orderWithPayment,
   billingInfo,
   shippingInfo,
   countries,
   selectedPaymentMethod,
   selectedShippingMethod,
 }: ICartSummaryPageProps) => {
-  const router = useRouter();
-  const theme = useTheme();
-  const { cart } = useCart();
-  const { t } = useTranslation();
-
-  const getCountryName = (countryId: string) =>
-    countries.find((c) => c.code == countryId)?.name || "";
-
-  const shippingInfoRows: ICartInfoTableRow[] = [
-    {
-      label: t("common:first-name-label"),
-      value: shippingInfo.first_name,
-    },
-    {
-      label: t("common:surname-label"),
-      value: shippingInfo.surname,
-    },
-    {
-      label: t("common:email-label"),
-      value: shippingInfo.email,
-    },
-    {
-      label: t("common:phone-label"),
-      value: shippingInfo.phone,
-    },
-    {
-      label: t("common:street-label"),
-      value: shippingInfo.street,
-    },
-    {
-      label: t("common:additional-info-label"),
-      value: shippingInfo.additional_info,
-    },
-    {
-      label: t("common:city-label"),
-      value: shippingInfo.city,
-    },
-    {
-      label: t("common:postal-code-label"),
-      value: shippingInfo.postal_code,
-    },
-    {
-      label: t("common:country-label"),
-      value: getCountryName(shippingInfo.country),
-    },
-  ];
-
-  const billingInfoRows: ICartInfoTableRow[] = [
-    {
-      label: t("common:first-name-label"),
-      value: billingInfo.first_name,
-    },
-    {
-      label: t("common:surname-label"),
-      value: billingInfo.surname,
-    },
-    {
-      label: t("common:company-name-label"),
-      value: billingInfo.company_name,
-    },
-    {
-      label: t("common:company-id-label"),
-      value: billingInfo.company_id,
-    },
-    {
-      label: t("common:vat-number-label"),
-      value: billingInfo.vat_number,
-    },
-    {
-      label: t("common:street-label"),
-      value: billingInfo.street,
-    },
-    {
-      label: t("common:additional-info-label"),
-      value: billingInfo.city,
-    },
-    {
-      label: t("common:postal-code-label"),
-      value: billingInfo.postal_code,
-    },
-    {
-      label: t("common:country-label"),
-      value: getCountryName(billingInfo.country),
-    },
-  ];
+  const order = orderWithPayment.order;
 
   return (
     <div className="container">
-      <CartStepper activeStep={3} />
-
-      <Grid container spacing={2}>
-        <Grid item xs={12} md={8}>
-          <Typography variant="h4" sx={{ my: 3 }}>
-            {t("cart:summary-title") /* Order summary */}
-          </Typography>
-          <CartItemList editable={false} />
-        </Grid>
-        <Grid item xs={12} md={4}>
-          <Box sx={{ [theme.breakpoints.up("md")]: { ml: 5 } }}>
-            <Typography variant="h6" sx={{ my: 3 }}>
-              {
-                t(
-                  "cart:shipping-and-payment-method-title"
-                ) /* Shipping and payment method */
-              }
-            </Typography>
-            <Table>
-              <CartMethodSummaryInfoRow
-                method={selectedShippingMethod.shipping_method}
-                formattedPrice={selectedShippingMethod.price_incl_vat}
-              />
-              <CartMethodSummaryInfoRow
-                method={selectedPaymentMethod.payment_method}
-                formattedPrice={selectedPaymentMethod.price_incl_vat}
-              />
-            </Table>
-          </Box>
-        </Grid>
-        <Grid item xs={12} md={4}>
-          <CollapsableContentWithTitle
-            title={t("cart:shipping-information-title")}
-          >
-            <CartInfoSummary rows={shippingInfoRows} />
-          </CollapsableContentWithTitle>
-        </Grid>
-        <Grid item xs={12} md={4}>
-          <CollapsableContentWithTitle title={t("billing-information-title")}>
-            <CartInfoSummary rows={billingInfoRows} />
-          </CollapsableContentWithTitle>
-        </Grid>
-
-        {cart ? (
-          <Grid item xs={12} md={8} textAlign="center">
-            <CartCompleteOrder cart={cart} />
-          </Grid>
-        ) : null}
-      </Grid>
-      <CartButtonRow
-        prev={{
-          title: t("back") /* Back */,
-          onClick: () => {
-            router.push("/cart/step/2");
-          },
-          disabled: false,
-        }}
+      <CartOrderSummary
+        cart={order.cart}
+        billingInfo={billingInfo}
+        shippingInfo={shippingInfo}
+        countries={countries}
+        selectedPaymentMethod={selectedPaymentMethod}
+        selectedShippingMethod={selectedShippingMethod}
+        creatingNewOrder={false}
       />
     </div>
   );
@@ -207,16 +62,16 @@ const CartSummaryPage = ({
  */
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const { req, res, locale } = context;
-  const { cartToken } = req.cookies;
+  const { orderId } = context.query;
 
-  if (cartToken === undefined || cartToken === null) {
-    return {
-      redirect: {
-        destination: "/",
-        permanent: false,
-      },
-    };
-  }
+  const order: IOrderWithPayment = await orderDetailAPI(
+    "GET",
+    orderId?.toString() || "",
+    req as NextApiRequest,
+    res as NextApiResponse
+  );
+
+  const cartToken = order.order.cart.token;
 
   const shippingInfo = await cartShippingInfoAPI(
     "GET",
@@ -254,6 +109,7 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 
   return {
     props: {
+      orderWithPayment: order,
       billingInfo,
       shippingInfo,
       countries,
@@ -268,4 +124,4 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
   };
 };
 
-export default CartSummaryPage;
+export default OrderDetailPage;
